requirements.txt
pandas
gspread
openai
oauth2client
tiktoken
faster-whisper
icecream
pyautogui
opencv-python
pyperclip
google-cloud-storage
pytest
fastapi
uvicorn
google-cloud-pubsub